+++ src/routes/routes.js
@@ -35,6 +35,8 @@ import PerfilDependente from "../pages/responsavel/dependentes/perfil-dependente
 import PerfilMotorista from "../pages/responsavel/dependentes/perfil-motorista/PerfilMotorista";
 import PerfilMotoristaSolitacao from "../pages/responsavel/dependentes/perfil-motorista/PerfilMotoristaSolicitacao";
 import RTempoReal from "../pages/responsavel/tempo_real/TempoReal";
+import RTempoRealMapa from "../pages/responsavel/tempo_real/TempoRealMapa";
+
 import Cad_Trajeto from "../pages/motorista/trajetos/Cad_Trajeto";
 
 function Rotas() {
@@ -156,6 +158,14 @@ function Rotas() {
           element={
             <PrivateRoute>
               <RTempoReal />
+            </PrivateRoute>
+          }
+        />
+        <Route
+          path="/responsavel/tempo-real/:id"
+          element={
+            <PrivateRoute>
+              <RTempoRealMapa />
             </PrivateRoute>
           }
         />
